+++ CHANGELOG.md
@@ -1,16 +1,26 @@
+## [1.1.1]
+
+### Fixed
+
+- Fixed a bug in the `NavigationView`, `PaneController` and `PaneItemDestination`.
+- Fixed memory leak `PaneController`.
+
 ## [1.1.0]
 
 ### Feat
+
 - Added `footers` property to Navigation Pane for displaying additional widgets after the main navigation items.
 - Added `additionalLending` feature to Navigation App Bar.
 
 ### Fixed
+
 - Fixed a bug in the `PaneController`.
 - Migrates deprecated `MaterialState` and `MaterialStateProperty` to `WidgetState` and `WidgetStateProperty`.
 
 ## [1.0.1]
 
 ### Fixed
+
 - Fixed a bug in the **Navigation App Bar** that affected layout in Compact mode.
 - Fixed a bug in the **Navigation Pane** that prevented correct operation in Compact mode.